pyproject.toml
[project]
name = "tequilatown-agent-swarm"
description = "TequilaTown Miami multi-agent AI operating system built on Agency Swarm and the OpenAI Agents SDK"
version = "0.1.0"
license = { text = "MIT" }
keywords = ["agency-swarm", "openai", "multi-agent", "tequilatown", "event-operations"]
requires-python = ">=3.12"
dependencies = [
    "agency-swarm[fastapi,jupyter,litellm]>=1.9.7",
    "questionary>=2.0.0",
    "python-dotenv",
    "rich",
    "fastapi",
    "uvicorn",
    "composio==0.8.0",
    "composio-openai-agents==0.8.0",
    "pytz",
    # Data analysis
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "statsmodels>=0.14.0",
    "scikit-learn>=1.3.0",
    # Visualisation
    "matplotlib>=3.5.0",
    "seaborn>=0.12.0",
    "plotly>=5.14.0",
    # Excel / CSV
    "openpyxl>=3.1.0",
    "xlrd>=2.0.0",
    # Jupyter / IPython
    "jupyter>=1.0.0",
    "ipython>=8.0.0",
    "nbformat>=5.0.0",
    # Image processing
    "pillow>=9.0.0",
    "opencv-python-headless",
    # PowerPoint / document processing
    "python-pptx>=1.0.0",
    "defusedxml>=0.7.0",
    "pdf2image>=1.16.0",
    "markitdown[pptx]>=0.1.0",
    "six>=1.16.0",
    "python-docx",
    "tinycss2",
    "beautifulsoup4",
    "cairosvg",
    "weasyprint",
    "html2text",
    "playwright",
    # AI / media generation
    "google-genai",
    "fal-client",
    "moviepy<2",
    "imageio-ffmpeg",
    "httpx",
]

[project.scripts]
openswarm = "run_utils:main"
tequilatown-agentswarm = "run_utils:main"

[tool.setuptools]
py-modules = ["agency", "swarm", "helpers", "config", "onboard", "server"]

[tool.setuptools.packages.find]
where = ["."]
exclude = ["agentswarm-cli*", "venv*", ".venv*", ".agency_swarm*", "node_modules*", "*.node_modules*", "activity*", "mnt*", "pptx*", "slides"]

[tool.setuptools.package-data]
"*" = ["*.md", "*.json", "agency-*"]
